fix: --longtest now working on toplevel
