+++ training/pytest.ini
@@ -3,5 +3,8 @@ markers =
     data_dependent: marks tests depending on data (deselect with '-m "not data_dependent"')
     auth: marks tests that require authentication (deselect with '-m "not auth"')
     gpu: marks tests that require a GPU (deselect with '-m "not gpu"')
+    longtest: mark test as long-running (skipped unless --longtests is used)
+
+addopts = --longtest
 
 tmp_path_retention_policy = none
