+++ training/tests/integration/test_training_cycle.py
@@ -18,12 +18,11 @@ from anemoi.training.train.train import AnemoiTrainer
 
 os.environ["ANEMOI_BASE_SEED"] = "42"  # need to set base seed if running on github runners
 
-longtests = pytest.mark.skipif("not config.getoption('longtests')", reason="need --longtests option to run")
 
 LOGGER = logging.getLogger(__name__)
 
 
-@longtests
+@pytest.mark.longtest
 def test_training_cycle_architecture_configs(architecture_config: DictConfig) -> None:
     AnemoiTrainer(architecture_config).train()
     shutil.rmtree(architecture_config.hardware.paths.output)
